+++ tweet-blog-public-vue3/src/config/info.ts
@@ -14,29 +14,29 @@ export const appInfo = {
     link: 'https://github.com/Tweblog/tweblog'
   },
   officialWebsite: {
-    text: 'tweblog.sakiko.top',
-    link: 'https://tweblog.sakiko.top/'
+    text: 'tweblog.pages.dev',
+    link: 'https://tweblog.pages.dev/'
   },
   officialDocs: {
     text: '网站 | 文档',
-    link: 'https://tweblog.sakiko.top/'
+    link: 'https://tweblog.pages.dev/'
   },
   importDocs: {
     text: '导入说明',
-    link: 'https://tweblog.sakiko.top/guide/feature/tweet-import'
+    link: 'https://tweblog.pages.dev/guide/feature/tweet-import'
   },
   // 删除导入记录
   importDeleteDocs: {
     text: '说明',
-    link: 'https://tweblog.sakiko.top/guide/feature/tweet-import#导入设置'
+    link: 'https://tweblog.pages.dev/guide/feature/tweet-import#导入设置'
   },
   proxyDocs: {
     text: '代理说明',
-    link: 'https://tweblog.sakiko.top/guide/feature/system-config#代理配置'
+    link: 'https://tweblog.pages.dev/guide/feature/system-config#代理配置'
   },
   // 转发配置
   forwardSettingDocs: {
     text: '配置说明',
-    link: 'https://tweblog.sakiko.top/guide/feature/tweet-forward'
+    link: 'https://tweblog.pages.dev/guide/feature/tweet-forward'
   }
 } as const